feat: move recherche to server component
