+++ components/a11y-components/hidden-h1/styles.module.scss
@@ -1,0 +1,10 @@
+.h1 {
+  clip: rect(1px, 1px, 1px, 1px);
+  height: 1px;
+  overflow: hidden;
+  position: absolute;
+  white-space: nowrap;
+  width: 1px;
+  z-index: -1;
+  user-select: none;
+}
